Move Projects into it's own PyDantic schema
