+++ src/app/schemas/projects.py
@@ -1,0 +1,28 @@
+from pydantic import BaseModel, ConfigDict, Field
+
+
+class ProjectModuleListResponse(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+    projects: list[str] = Field(description="Registered project module names")
+
+
+class ProjectModuleContractStatus(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+    project_module: str = Field(description="Project module identifier")
+    valid: bool = Field(description="Whether discovery contract validation passed")
+    required_adapters: list[str] = Field(default_factory=list)
+    error: str | None = Field(default=None, description="Validation/import error when invalid")
+    exports: list[str] = Field(
+        default_factory=list,
+        description="Known module exports relevant to discovery integration",
+    )
+    enrichment_keys: list[str] = Field(
+        default_factory=list,
+        description="Discover bundle enrichment keys this module uses (e.g. ra_dec_vsys, sbid_to_eval_file)",
+    )
+
+
+class ProjectModuleContractListResponse(BaseModel):
+    model_config = ConfigDict(extra="forbid")
+    count: int
+    modules: list[ProjectModuleContractStatus]
